remove MainActivity, move utils to util package, update kotlin version
